Group senior statements into sections (#9)
I've used the same structure as for mid-level, in an attempt to
distinguish between responsibilities/behaviours and skills/knowledge.

This adds back some of the structure that was present in the original
document, but this is just a first step. I think this section could
still do with a lot of editing.

This role has more statements than any other in the technologist's
career path and I think we could more clearly distinguish what the core
responsibilities and skills actually are.

Part of the problem is there are a lot of responsibilities mixed in that
don't apply to every senior; for example:

  - acting as a technical lead or line manager

  - go to expert on one area of the codebase

  - statements relating to civil service corporate objectives, which can
    be met in any number of ways ("visible externally", "helping the
wider civil service", "wider developer community, blogging or speaking
at conferences", "helping colleagues with career development"). I think
we could cover these things (they are duplicated elsewhere) but they
don't feel very role-specific.

Also, in the original document, the first few statements are more of a
summary, and there is some duplication between these and statements
further down so there we could perhaps combine some of the competencies.