Link to GDS tech manual
We now have https://github.com/alphagov/gds-tech which documents how we
do stuff at GDS, including particular technologies and software we've
chosen to use.

I think some of this gives specific examples of things we talk about in
relation to competencies, and we can tie it all together.

I've added a section to the contributing guide about what should go in a
guide here and when we should link out to other resources.

There's room to expand this, but hopefully this is a good starting
point.

https://github.com/alphagov/gds-tech-learning-pathway/issues/18