Add senior and tech lead to leading on stories
I've shoehorned tech leading in here, because the current
career path doesn't say much about this, except "you can be a tech lead"

This is taken from the verify team manual, which describes the role
very well. The main part I've left out is pastoral care of developers,
because I think this is a separate competency that is also relevant
to anyone doing line management.

The main takeaway should be that tech leading at gds is a natural
progression from leading on stories at mid level.