Remove dead link and structure the junior page
I've split up the production-ready-code competency, so we should link
to working-as-part-of-a-team and designing-for-reliability instead.

I've taken the opportunity to divide this up into sections.
I think this will help to communicate what we expect juniors to be doing
and why. The technical stuff links to resource pages, whereas the rest
of it has a competency page. I think this works ok because the techy
stuff is more opinionated and people's learning needs will vary
depending on what they're working on.