Update contributing guide and readme
The contributing guide felt a bit too prescriptive and people
are already contributing in other ways that are really helpful,
so encourage other ways of contributing.

I've also taken the opportunity to update the inspiration page
and link to it from the readme, and update the status in the readme.