Replace individual role pages with a single career path page
The individual pages were often duplicating each other and it was
hard to join them up. It's hard to make sense of a list of competencies
without comparing them to another role.

We've now come up with a common list of competencies that is represented
throughout the career path, and our template has a section for each
level (junior - senior). So I think we just need a single list of
what they are so you can drill down the areas you're interested in.

It's still a bit difficult to get an idea of "what am I missing to get
promoted" but I think this structure will be a better foundation,
as it lets people drill down into the competencies they're least
confident about or most interested in developing.

I've also used this page to talk about the learning opportunities
available more generally, in particular to call out the 5 days
of protected time we're entitled to as civil servants.

This mainly comes from this document from 2016:
https://docs.google.com/document/d/1dVdr6djiKizco7cMXmQj25NW7DSUODZ2LxftDFxl5lk/edit