Edit the code review page
This is quite a big edit that aims to reduce the amount of information
overload on this page, and avoid overlap with the GDS way.

In particular I've stripped down the section on how to give a code
review, and instead just emphasised one skill (being nice to people)
and deferred to the GDS way for the actual process of giving a code
review and what things to look for when reviewing software at GDS.

However, between the GDS way and the learning pathways, we've collected
a lot of blog posts and external guides about code review, so I've
included these in a section at the end. These are useful for anyone
seeking out different perspectives than the GDS one. In particular,
the technology community encompasses other disciplines such as data
science and technical writing, who can also benefit from code review
but come at it from a different perspective than software develoeprs.

I've also expanded the section on "ways to streamline your code review"
to include things that were previously mentioned in other sections.
This section is not really GDS specific and is "take it or leave it"
advice that showcases things some teams have found useful.